AG-17444 Relocate detached-attach regression tests
Move the AG-17444 regression coverage out of the standalone
chartTypeChangeCanvas.test.ts into two more fitting homes:

- A DOMManager.destroy() unit test in domManager.test.ts that pins the
  ownership check directly: a re-homed (transferred) child survives
  teardown while a still-owned child is removed.
- An AgCharts.create() -> update(polar) integration test in
  agChart.test.ts, alongside the other create/update integration tests.

Both fail without the destroy() fix; the cartesian (line) control case
keeps passing, matching the type-change-only nature of the bug. No
production code changed.
