+++ packages/ag-charts-community/src/chart/agChart.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, test } from 'vitest';
 
-import type { AgChartInstance } from 'ag-charts-types';
+import type { AgChartInstance, AgChartOptions } from 'ag-charts-types';
 
 import { AgCharts } from '../api/agCharts';
 import { NumberAxis } from './axis/numberAxis';
@@ -8,7 +8,7 @@ import { themes } from './mapping/themes';
 import { AreaSeries } from './series/cartesian/areaSeries';
 import { BarSeries } from './series/cartesian/barSeries';
 import { LineSeries } from './series/cartesian/lineSeries';
-import { Chart, deproxy, setupMockCanvas, setupMockConsole, waitForChartStability } from './test/utils';
+import { Chart, delay, deproxy, setupMockCanvas, setupMockConsole, waitForChartStability } from './test/utils';
 
 const revenueProfitData = [
     {
@@ -568,4 +568,51 @@ describe('AgChart', () => {
             },
         });
     });
+});
+
+describe('Chart re-creation canvas attachment (AG-17444)', () => {
+    setupMockCanvas();
+
+    let chart: AgChartInstance | undefined;
+    afterEach(() => {
+        chart?.destroy();
+        chart = undefined;
+    });
+
+    const DATA = [
+        { k: 'A', v: 1 },
+        { k: 'B', v: 2 },
+        { k: 'C', v: 3 },
+    ];
+
+    const SERIES: Record<string, NonNullable<AgChartOptions['series']>> = {
+        // Polar series change the chart type from the default cartesian chart, which re-creates
+        // the chart and transfers the scene (the AG-17444 / AS-774 path).
+        donut: [{ type: 'donut', angleKey: 'v', calloutLabelKey: 'k' }],
+        pie: [{ type: 'pie', angleKey: 'v', calloutLabelKey: 'k' }],
+        // Cartesian control: no type change, no re-creation.
+        line: [{ type: 'line', xKey: 'k', yKey: 'v' }],
+    };
+
+    // Regression: `AgCharts.create()` followed immediately by an `update()` that changes the
+    // chart type transfers the scene to a re-created chart. The previous chart's deferred
+    // teardown (DOMManager.destroy) must not remove the transferred <canvas> that the new chart
+    // has already adopted — otherwise the live canvas is orphaned from the DOM and renders blank.
+    test.each(Object.keys(SERIES))(
+        'canvas stays attached to the container when update() applies a %s series',
+        async (seriesType) => {
+            const container = document.createElement('div');
+            document.body.appendChild(container);
+
+            chart = AgCharts.create({ container } as AgChartOptions);
+            await chart.update({ container, data: DATA, series: SERIES[seriesType] } as AgChartOptions);
+            await waitForChartStability(chart);
+            // Allow the re-created chart's deferred teardown to run before asserting.
+            await delay(50);
+
+            const canvasElement = deproxy(chart).ctx.scene.canvas.element;
+            expect(canvasElement.isConnected).toBe(true);
+            expect(container.contains(canvasElement)).toBe(true);
+        }
+    );
 });
